backend/requirements.prod.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.4; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-extensions==3.2.3; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
